requirements.txt
cryptography>=2.9.2
python-dotenv>=0.15.0
pycrypto>=2.6.1
pycryptodome>=3.10.1